requirements.txt
llama-cpp-python==0.3.14
pydantic==2.10.6 
gradio==5.21.0

